Well define the code structure [OK]
Prepare a raw test map         [OK]

Implement the minimap          [DOING]
  Fill the image               [TODO]
    Floor cells                [TODO]
    Wall cells                 [TODO]
    Player positon             [TODO]
    Player direction           [TODO]
    Player field of view       [TODO]

+++ private/config.h
@@ -6,7 +6,7 @@
 /*   By: jodufour <jodufour@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/05/12 15:10:29 by jodufour          #+#    #+#             */
-/*   Updated: 2023/05/14 01:08:50 by jodufour         ###   ########.fr       */
+/*   Updated: 2023/05/14 23:43:08 by jodufour         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,22 +14,22 @@
 # define CONFIG_H
 
 # include <stdbool.h>
-# include <stddef.h>
+# include <stdint.h>
 
 /* ************************************************************************** */
 /*                                   Window                                   */
 /* ************************************************************************** */
-static bool const	g_full_screen = false;
+static bool const		g_full_screen = false;
 
-static size_t const	g_window_width = 1000LU;
-static size_t const	g_window_height = 600LU;
+static uint16_t const	g_window_width = 1000U;
+static uint16_t const	g_window_height = 600U;
 
-static char *const	g_window_title = "cub3D";
+static char *const		g_window_title = "cub3D";
 
 /* ************************************************************************* */
 /*                                  Minimap                                  */
 /* ************************************************************************* */
-static size_t const	g_minimap_width = 420LU;
-static size_t const	g_minimap_height = 210LU;
+static uint16_t const	g_minimap_width = 420U;
+static uint16_t const	g_minimap_height = 210U;
 
 #endif
